fix: fixed changed from int to float read
huge save could have been very bad

+++ race/stdinInput.js
@@ -11,7 +11,7 @@ export class StdinInput {
     }
 
     async waitInput() {
-        return (await this.it.next()).value.trim().split(" ").map((i) => parseInt(i));
+        return (await this.it.next()).value.trim().split(" ").map((i) => parseFloat(i));
     }
 
     output(data) {
